fix: harden channel property handling
